src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
bcrypt==4.0.1
    # via paramiko
blosc2==2.0.0
    # via tables
boto3==1.26.150
    # via -r requirements.in
botocore==1.29.150
    # via
    #   boto3
    #   s3transfer
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
contourpy==1.0.7
    # via matplotlib
coverage==7.2.7
    # via -r requirements.in
cycler==0.11.0
    # via matplotlib
cython==0.29.35
    # via
    #   -r requirements.in
    #   tables
decorator==5.1.1
    # via jsonpath-ng
exchange-calendars==4.2.8
    # via pandas-market-calendars
fonttools==4.39.4
    # via matplotlib
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via -r requirements.in
idna==3.4
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonpath-ng==1.5.3
    # via -r requirements.in
kiwisolver==1.4.4
    # via matplotlib
korean-lunar-calendar==0.3.1
    # via exchange-calendars
lz4==4.3.2
    # via -r requirements.in
matplotlib==3.7.1
    # via seaborn
msgpack==1.0.5
    # via blosc2
numexpr==2.8.4
    # via tables
numpy==1.24.3
    # via
    #   -r requirements.in
    #   contourpy
    #   exchange-calendars
    #   matplotlib
    #   numexpr
    #   pandas
    #   pyarrow
    #   seaborn
    #   tables
packaging==23.1
    # via
    #   matplotlib
    #   tables
pandas==2.0.2
    # via
    #   -r requirements.in
    #   exchange-calendars
    #   pandas-market-calendars
    #   seaborn
pandas-market-calendars==4.1.4
    # via -r requirements.in
paramiko==3.1.0
    # via pysftp
pbr==5.11.1
    # via testresources
pillow==9.5.0
    # via matplotlib
ply==3.11
    # via jsonpath-ng
py-cpuinfo==9.0.0
    # via tables
pyarrow==12.0.0
    # via -r requirements.in
pyluach==2.2.0
    # via exchange-calendars
pyparsing==3.0.9
    # via matplotlib
pypubsub==4.0.3
    # via -r requirements.in
pysftp==0.2.9
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   exchange-calendars
    #   matplotlib
    #   pandas
    #   pandas-market-calendars
pytz==2023.3
    # via
    #   exchange-calendars
    #   pandas
    #   pandas-market-calendars
requests==2.31.0
    # via -r requirements.in
s3transfer==0.6.1
    # via boto3
seaborn==0.12.2
    # via -r requirements.in
six==1.16.0
    # via
    #   jsonpath-ng
    #   python-dateutil
smmap==5.0.0
    # via gitdb
tables==3.8.0
    # via -r requirements.in
testresources==2.0.1
    # via -r requirements.in
toolz==0.12.0
    # via exchange-calendars
tzdata==2023.3
    # via pandas
urllib3==1.26.16
    # via
    #   botocore
    #   requests
websockets==11.0.3
    # via -r requirements.in
